requirements.txt
Flask==3.1.0
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
Pillow==11.0.0
pygame==2.6.1
Requests==2.32.3
torch==2.4.1+cu118
torchvision==0.19.1+cu118
ultralytics==8.3.4
